+++ src/telegram_llm_bot/bots/base_chatbot/settings.py
@@ -9,6 +9,7 @@ from telegram_llm_bot.paths import load_environment
 load_environment()
 
 from telegram_llm_bot.bots.base_chatbot.handlers.text import text_chat_handler
+from telegram_llm_bot.bots.base_chatbot.handlers.voice import note_handler
 
 logger = logging.getLogger(__name__)
 
@@ -18,6 +19,7 @@ class Settings:
     def __init__(self, **values: Any):
         self.handlers: List[BaseHandler] = [
             MessageHandler(filters.TEXT, text_chat_handler, block=False),
+            MessageHandler(filters.VOICE, note_handler, block=False),
         ]
         self.commands: Dict[str, str] = {}
         config = load_bot_config()